moved ScheduleForm children components to separate files/directories
