+++ applications/launchpad/gui-react/src/containers/MiningContainer/Scheduling/ScheduleForm/DateScheduler/styles.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+export const HumanReadableScheduledDate = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: 'center';
+  margin-top: ${({ theme }) => theme.spacing()};
+`
