+++ applications/launchpad/gui-react/src/containers/MiningContainer/Scheduling/ScheduleForm/WeekdaySelector/styles.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+export const Weekday = styled.div`
+  width: 48px;
+  height: 48px;
+  background-color: ${({ theme }) => theme.backgroundImage};
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border-radius: 4px;
+  cursor: pointer;
+`
