+++ applications/launchpad/gui-react/src/containers/MiningContainer/Scheduling/index.tsx
@@ -28,12 +28,12 @@ const SchedulingContainer = ({
   onClose: () => void
 }) => {
   const [idToEdit, setScheduleToEdit] = useState('')
-  const [editOpen, setEditOpen] = useState(false)
+  const [editOpen, setEditOpen] = useState(true)
   const schedules = useAppSelector(selectSchedules)
   const scheduleToEdit = useAppSelector(selectSchedule(idToEdit))
   const dispatch = useAppDispatch()
 
-  const clearEditState = () => {
+  const stopEditing = () => {
     setEditOpen(false)
     setScheduleToEdit('')
   }
@@ -49,14 +49,14 @@ const SchedulingContainer = ({
   }
 
   const close = () => {
-    clearEditState()
+    stopEditing()
     onClose()
   }
 
   return (
     <Modal open={open} onClose={close} size='small'>
       <ScheduleContainer>
-        {false && !editOpen && (
+        {!editOpen && (
           <ScheduleList
             schedules={schedules}
             cancel={close}
@@ -66,9 +66,10 @@ const SchedulingContainer = ({
             remove={scheduleId => dispatch(removeSchedule(scheduleId))}
           />
         )}
-        {(editOpen || true) && (
+        {editOpen && (
           <ScheduleForm
             value={scheduleToEdit}
+            cancel={stopEditing}
             onChange={value =>
               dispatch(updateSchedule({ value, scheduleId: idToEdit }))
             }
